add: cloudinary dependency, cloudinary connection, uploadImage. change: types to custom.d.ts
