+++ controlers/mentor.ts
@@ -1,22 +1,13 @@
+import { uploadImgToCloudinary } from "lib/cloudinary";
 import { Mentor } from "models/mentors";
 
-type MentorData = {
-    name: string,
-    category:string,
-    community:string,
-    description:string,
-    image: string
+export async function createNewMentor(data: MentorData) {
+  try {
+    const image = await uploadImgToCloudinary(data.image);
+    const mentor = { ...data, image };
+    const result = await Mentor.createNewMentor(mentor);
+    return result;
+  } catch (error) {
+    throw error;
+  }
 }
-
-export async function createNewMentor(data:MentorData) {
-    try {
-        
-       const result = await  Mentor.createNewMentor(data)
-
-       return result
-        
-
-    } catch (error) {
-        throw error
-    }
-  }
