+++ custom.d.ts
@@ -1,0 +1,7 @@
+type MentorData = {
+  name: string;
+  category: string;
+  community: string;
+  description: string;
+  image: string;
+};
