+++ package.json
@@ -13,6 +13,7 @@
     "@types/node": "18.11.10",
     "@types/react": "18.0.26",
     "@types/react-dom": "18.0.9",
+    "cloudinary": "^1.32.0",
     "cors": "^2.8.5",
     "eslint": "8.29.0",
     "eslint-config-next": "13.0.6",
